+++ core/async-operation/src/operation.ts
@@ -1,1 +1,2 @@
 export * from "./lib/AsyncOperation.js";
+export * from "./lib/AsyncOperationSubscriber.js";