+++ core/async-operation/test/async-operation.recipes.ts
@@ -1,4 +1,4 @@
-import { AsyncOperation } from "../src/operation.js";
+import { AsyncOperation, AsyncOperationSubscriber } from "../src/operation.js";
 
 await using operation = new AsyncOperation<string, number>(async ({ signal, write }) => {
 	if (signal.aborted) {
@@ -16,3 +16,19 @@ for await (const value of operation) {
 }
 
 console.log(await operation.result);
+
+await using subscriber = new AsyncOperationSubscriber<number, string>();
+using _evenValues = subscriber
+	.filter((value) => value % 2 === 0)
+	.subscribe((value) => {
+		console.log(value);
+	});
+
+await subscriber.consume(
+	new AsyncOperation<number, string>(async ({ write }) => {
+		await write(1);
+		await write(2);
+
+		return "complete";
+	}),
+);