+++ lit/signals/src/directives/operate.ts
@@ -1,3 +1,0 @@
-import { AsyncOperation } from "@serve-tools/async-operation";
-
-export { AsyncOperation };